Fixed Issue #2, implemented proper callback methods for requestControl.
Migrated onDeleteEvents and onDeleteStations that was used to change
the label from the events and stations control buttons to the new
schemma. Now we can register as many callbacks as necessary to each
callback name (like the mapping.js).

Also, when the user adds events -> Relative Mode, when the user delete events
-> Absolute Mode. Reset button on Submit also chooses the best based on the
existence of events in the requestControl.

Closes Issue #2.