New version of testEvent.py sent by Peter per mail (07.03.2014) to Massimo.
The text of the mail was:
I've written some more unit tests in testEvent.py to
call the service at  http://eida.rm.ingv.it/webinterface/wsgi/event/ingv
directly. It checks:

 - calling with no parameters

 - calling with format=json

 - the limit restriction (doesn't work with format=json, hmm)

 - latitude restrictions: maxlat and minlat work

 - longitude restrictions: maxlon and minlon work

 - depth restriction: setting maxdepth/mindepth seems
   to work, but I can NOT select deep/shallow events using
   the slider in the web page, for the ingv catalog.
   e.g. events >200 km.

These are in my modified testEvent.py, in which
I've disabled the tests of the other services (around line 1266),
for now. You're welcome to extend the tests defined in
class TestEventServiceFDSN.

P.