Add all the main improvements in the python files from the branch issue4.
We had to do this before merge becuase the python files suffered a lot of
minor changes related to formating and a merge would have resulted in a lot
of conflicts. The next step is to do the merge for all the files. In the case
of these python files we will accept always the version from the master branch,
as it is currently updated.